Research & Analysis Agents/Academic paper explainer agent/requirements.txt
streamlit
openai
python-dotenv
pandas
PyPDF2
